+++ init.py
@@ -5,7 +5,7 @@ import discord
 from discord.ext import commands
 from pretty_help import PrettyHelp
     
-client = commands.Bot(command_prefix = ':', help_command = PrettyHelp(dm_help = True))
+client = commands.Bot(command_prefix = ':', help_command = PrettyHelp())
 
 @client.command(help = 'Loads package.')
 @commands.is_owner()